Backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
scikit-learn==1.5.2 
rank-bm25==0.2.2
sentence-transformers==3.1.1
langchain==0.2.16
langchain-community==0.2.16
langchain-pinecone==0.1.3
pinecone-client[grpc]==5.0.1
httpx==0.27.2
structlog==24.4.0
reportlab==4.2.2
python-dotenv==1.0.1 
google-generativeai==0.7.2
